MAV-1382/tested api function
